using testgen mode
